refactor live reload injection stream
